Fix the health stats, add a defense stat
The damage formula is just damage = skillPower/defense
It's a *placeholder* !

+++ src/mob.h
@@ -33,7 +33,7 @@ private:
 
 public:
     MobType type;
-    unsigned health;
+    float health;
 
 private:
     Mobzone* spawnZone;
